+++ packages/nitro/example/lib/src/math.g.dart
@@ -48,17 +48,13 @@ class _MathImpl extends Math {
   late final void Function(int) _releaseUpdatesPtr =
       _dylib.lookupFunction<Void Function(Int64), void Function(int)>(
           'math_release_updates_stream');
-  // ignore: unused_field
   late final Pointer<NitroErrorFfi> Function() _getErrorPtr =
       _dylib.lookupFunction<Pointer<NitroErrorFfi> Function(),
           Pointer<NitroErrorFfi> Function()>('math_get_error');
-  // ignore: unused_field
   late final void Function() _clearErrorPtr = _dylib
       .lookupFunction<Void Function(), void Function()>('math_clear_error');
-  // ignore: unused_field
   late final Pointer<NativeFunction<Pointer<NitroErrorFfi> Function()>>
       _getErrorNativePtr = _dylib.lookup('math_get_error');
-  // ignore: unused_field
   late final Pointer<NativeFunction<Void Function()>> _clearErrorNativePtr =
       _dylib.lookup('math_clear_error');